requirements.txt
aniso8601==7.0.0
certifi==2019.6.16
Click==7.0
Flask==1.1.1
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
pytz==2019.1
six==1.12.0
Werkzeug==0.15.5
